Update README.md
Sudoku is a logic-based puzzle that uses combinatorial-number placement. In a classic Sudoku puzzle, the task is to fill the numbers in a 9 x 9 grid laced in such a way that each row, each column, and each of the sub-grid of the size 3 x 3 contains all the numbers from 1 to 9 (1 and 9 inclusive).